requirements.txt
streamlit>=1.44.1
google-cloud-vision>=3.10.1
google-cloud-storage>=2.19.0
google-cloud-aiplatform>=1.90.0
requests==2.31.0
beautifulsoup4>=4.13.4
plotly>=6.0.1
pillow>=11.2.1
seaborn>=0.13.2
matplotlib>=3.10.1
orjson>=3.10.17
statsmodels>=0.14.4
# altair>=5.5.0
# altair-viewer>=0.4.0